requirements.txt
pyautogui
selenium
webdriver_manager
Pillow
opencv-python
numpy
pywin32
customtkinter
